+++ apps/backend/src/main/java/com/sait/peelin/model/Customer.java
@@ -28,9 +28,8 @@ public class Customer {
     @JoinColumn(name = "user_id")
     private User user;
 
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "address_id", nullable = false)
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "address_id")
     private Address address;
 
     @NotNull